+++ eox_nelp/pearson_vue_engine/admin.py
@@ -9,7 +9,7 @@ in the Django admin interface.
 from django.contrib import admin
 from django.contrib.auth import get_user_model
 
-from eox_nelp.pearson_vue.models import PearsonEngine
+from eox_nelp.pearson_vue_engine.models import PearsonEngine
 
 User = get_user_model()
